seperated search comp in extention
